Refactor evolution middleware Fix minor bugs
