+++ src/components/GridContainer/GridContainer.js
@@ -4,6 +4,7 @@ const GridContainer = styled.div`
   background-color: ${({ theme }) => theme.surface.regular};
   border-radius: 14px;
   padding: 20px 20px 30px;
+  min-width: 0;
 `
 
 export default GridContainer
