+++ src/evolution/middleware/index.js
@@ -1,0 +1,19 @@
+import worker from './worker'
+import done from './done'
+import initialEvolution from './initialEvolution'
+import pause from './pause'
+import play from './play'
+import predictions from './predictions'
+import progress from './progress'
+import reset from './reset'
+
+export default [
+  worker,
+  done,
+  initialEvolution,
+  pause,
+  play,
+  predictions,
+  progress,
+  reset,
+]
