+++ src/evolution/worker/middleware/index.js
@@ -1,0 +1,19 @@
+import progress from './progress'
+import instance from './instance'
+import initialEvolution from './initialEvolution'
+import predictions from './predictions'
+import reset from './reset'
+import start from './start'
+import step from './step'
+import done from './done'
+
+export default [
+  progress,
+  instance,
+  initialEvolution,
+  predictions,
+  reset,
+  start,
+  step,
+  done,
+]
