+++ src/evolution/worker/utils.js
@@ -1,11 +1,20 @@
 /* eslint-env worker */
 import { evolutionProgress } from '../actions'
 
-export const handleNewMessagesFirst = async (func, ...args) =>
+/**
+ * Interupts the event loop to handle other events first.
+ *
+ * @async
+ * @function interupt
+ * @param {function} func - The function to be called when the event loop continues.
+ * @param  {...any} args - The arguments for the function `func`.
+ * @return {Promise<any>} The return of the function `func` called with arguments `args`.
+ */
+export const interupt = async (func, ...args) =>
   new Promise(resolve => {
     setTimeout(() => {
       resolve(func(...args))
-    }, 0)
+    }, 1)
   })
 
 export const createExpression = (
@@ -81,7 +90,7 @@ export const loop = async (store, action) => {
     const progressAction = evolutionProgress(result)
     progressAction.meta = { throttle: true }
 
-    await handleNewMessagesFirst(store.dispatch, progressAction)
+    await interupt(store.dispatch, progressAction)
 
     if (!checkIfEvolving(store)) {
       break
