+++ src/utils/throttle.js
@@ -1,4 +1,13 @@
-export default function throttle(func, limit) {
+/**
+ * Creates a throttled version of the `func`.
+ * `func` will only be executed with a minimal interval of `limit`.
+ * Subsequent calls to the generated function faster then `limit` will have no effect.
+ * @function throttle
+ * @param {function} func - The function to be throttled.
+ * @param {number} limit - The minimum interval in milliseconds `func` needs to wait before it can be called again.
+ * @returns {function} The throtteled version of `func`.
+ */
+const throttle = (func, limit) => {
   let inThrottle
 
   return (...args) => {
@@ -13,3 +22,5 @@ export default function throttle(func, limit) {
     }
   }
 }
+
+export default throttle
